# public/sample-specs/vampi-openapi3.yml
openapi: 3.0.0
info:
  title: VAmPI - Vulnerable API
  version: 1.0.0
  description: 취약한 API 예제 (OWASP API Top 10 학습용)
servers:
  - url: http://localhost:5000
    description: 로컬 개발 서버
paths:
  /api/users/{userId}:
    get:
      summary: 사용자 정보 조회
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
                  password:
                    type: string
                  creditCard:
                    type: string
      # 보안 설정 없음 - BOLA 취약점
      
    put:
      summary: 사용자 정보 수정
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: 성공
      # 인증/인가 없음 - Mass Assignment 취약점
...